% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emodnet_wcs.R
\name{emdn_wcs}
\alias{emdn_wcs}
\title{Which data sources (services) are available?}
\usage{
emdn_wcs()
}
\value{
Tibble of available EMODnet Web Coverage Services
}
\description{
Available EMODnet Web Coverage Services
}
\examples{
emdn_wcs()

}
